+++ lib/core-complete.sh
@@ -3115,7 +3115,7 @@ function ble/complete/progcomp/.compgen-helper-func {
   local cmd=${COMP_WORDS[0]} cur=${COMP_WORDS[COMP_CWORD]} prev=${COMP_WORDS[COMP_CWORD-1]}
   ble/function#push compopt 'ble/complete/progcomp/compopt "$@"'
 
-  # WA for blocking scp/ssh #D1807
+  # WA (#D1807): A workaround for blocking scp/ssh
   ble/function#push ssh '
     local IFS=$_ble_term_IFS
     if [[ " ${FUNCNAME[*]} " == *" ble/complete/progcomp/.compgen "* ]]; then
@@ -3126,8 +3126,13 @@ function ble/complete/progcomp/.compgen-helper-func {
       ble/function#push/call-top "$@"
     fi'
 
+  # WA (#D1834): Suppress invocation of "command_not_found_handle" in the
+  #   completion functions
+  ble/function#push command_not_found_handle
+
   builtin eval '"$comp_func" "$cmd" "$cur" "$prev"' < /dev/null >&$_ble_util_fd_stdout 2>&$_ble_util_fd_stderr; local ret=$?
 
+  ble/function#pop command_not_found_handle
   ble/function#pop ssh
   ble/function#pop compopt