+++ docs/ChangeLog.md
@@ -362,6 +362,7 @@
 - progcomp: work around the cobra V2 description hack (reported by SuperSandro2000) `#D1803` 71d0736
 - complete: work around blocking `_scp_remote_files` and `_dnf` (reported by iantra) `#D1807` a4a779e 46f5c13
 - history: work around broken timestamps in `HISTFILE` `#D1831` XXXXXXX
+- progcomp: disable `command_not_found_handle` `#D1834` XXXXXXX
 
 ## Internal changes and fixes